requirements.txt

dvc[gs]==3.30.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.35.0
mlflow==2.7.1
feast==0.34.0
pytest==7.4.2

